+++ app/src/main/java/org/codeforbirmingham/spotholes/ui/PotholeListFragment.java
@@ -1,4 +1,4 @@
-package  org.codeforbirmingham.potholes.ui;
+package  org.codeforbirmingham.spotholes.ui;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
@@ -8,7 +8,7 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import org.codeforbirmingham.potholes.R;
+import org.codeforbirmingham.spotholes.R;
 
 public class PotholeListFragment extends Fragment {
