+++ app/src/main/java/org/codeforbirmingham/spotholes/ui/widget/DividerItemDecoration.java
@@ -1,4 +1,4 @@
-package org.codeforbirmingham.potholes.ui.widget;
+package org.codeforbirmingham.spotholes.ui.widget;
 
 import android.content.Context;
 import android.content.res.TypedArray;
