+++ app/src/main/java/org/codeforbirmingham/spotholes/utils/Util.java
@@ -1,4 +1,4 @@
-package org.codeforbirmingham.potholes.utils;
+package org.codeforbirmingham.spotholes.utils;
 
 import android.content.Context;
 import android.util.DisplayMetrics;
